Added passive infrared motion detector support to the Monitron

+++ monitron/libraries/AnalogSampler/AnalogSampler.h
@@ -28,7 +28,7 @@ class AnalogSampler
     double _maxValue;
     double _sumOfValues;
     double _sumOfSquares;
-    unsigned int _n;
+    unsigned long _n;
 };
 
 #endif
